sync using `UpdatesByRange`
The light client is now synced with `UpdatesByRange` until it has access
to the current sync committee. It will also schedule a timer to fetch
the next sync committee so that it is ready when needed. Note that the
light client does not sync the latest finalized / optimistic header at
this time. Furthermore, the UI is extended to show optimistic header.