fetch `transactionsRoot` and `withdrawalsRoot` from EL
Use execution-API to obtain `transactionsRoot` and `withdrawalsRoot`.

Note that at this time, ELs don't always return `withdrawalsRoot` in
response to `eth_getBlockByHash` and `eth_getBlockById`.

Also note that in order to verify inclusion of a `transaction`, all
transactions from the block must be downloaded. There is no proof API,
and the transaction hashes cannot be used to compute `transactionsRoot`.