Remove Merkleization illustrations
The Merkleization of `StableContainer` is better understood with many
implementations having made progress towards https://stabilitynow.box.
As the level of detail of inner containers is not visualized and the
signature is simplified to no longer have the `from` address, the
illustrations on the SSZ page is no longer useful enough to warrant
continuous maintenance, they also don't render properly on mobile.